Move Scope struct to its own file, making the frame module private in the process.
