+++ vm/src/lib.rs
@@ -56,7 +56,7 @@ mod dictdatatype;
 pub mod eval;
 mod exceptions;
 pub mod format;
-pub mod frame;
+mod frame;
 mod frozen;
 pub mod function;
 pub mod import;
@@ -64,6 +64,7 @@ pub mod obj;
 pub mod py_serde;
 mod pyhash;
 pub mod pyobject;
+pub mod scope;
 pub mod stdlib;
 mod sysmodule;
 mod traceback;
